+++ app/src/main/java/com/example/calculus/MainActivity.java
@@ -6,8 +6,11 @@ import androidx.appcompat.app.AppCompatDelegate;
 import androidx.appcompat.widget.SwitchCompat;
 
 import android.content.Context;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
 import android.widget.CompoundButton;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -16,7 +19,7 @@ public class MainActivity extends AppCompatActivity {
 
     private TextView displayTask;
     private TextView displayResult;
-    private SwitchCompat nightSwitch;
+    private Button buttonTheme;
     private String task = "";
     private String result = "";
     private final String[] ACTIONS = new String[] {"+", "-", "x", "÷"};
@@ -32,30 +35,23 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        sharedPreferences = getSharedPreferences(MY_PREFERENCES, Context.MODE_PRIVATE);
 
-        nightSwitch = findViewById(R.id.switchNightMode);
 
-        NightModeIsActivated();
+
 
         displayTask = findViewById(R.id.display);
         displayResult = findViewById(R.id.result);
 
-        nightSwitch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+
+
+        buttonTheme = findViewById(R.id.b_theme);
+        buttonTheme.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
-                if (b) {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
-                    saveNightModeState(true);
-                    recreate();
-                } else {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
-                    saveNightModeState(false);
-                    recreate();
-                }
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, ThemeChooser.class);
+                startActivity(intent);
             }
         });
-
 
         findViewById(R.id.b_one).setOnClickListener(view -> {
             task = task + "1";
@@ -112,7 +108,7 @@ public class MainActivity extends AppCompatActivity {
                 if (task.isEmpty()){
                     task = task + "0.";
                 } else if (task.contains(".")){
-                    Toast.makeText(getApplicationContext(), "Your number already has a dot", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(getApplicationContext(), R.string.one_dot, Toast.LENGTH_SHORT).show();
                 } else {
                     task = task + ".";
                 }
@@ -136,7 +132,7 @@ public class MainActivity extends AppCompatActivity {
                     (task.charAt(task.length() - 1) == '.')){
                 task = task.substring(0, task.length() - 1) + "+";
             } else if (hasActionSign(task)){
-                Toast.makeText(getApplicationContext(), "Only one action a time, sorry", Toast.LENGTH_SHORT).show();
+                Toast.makeText(getApplicationContext(), R.string.one_action, Toast.LENGTH_SHORT).show();
             } else {
                 task = task + "+";
             }
@@ -152,7 +148,7 @@ public class MainActivity extends AppCompatActivity {
                     (task.charAt(task.length() - 1) == '.')) {
                 task = task.substring(0, task.length() - 1) + "-";
             } else if (hasActionSign(task)){
-                Toast.makeText(getApplicationContext(), "Only one action a time, sorry", Toast.LENGTH_SHORT).show();
+                Toast.makeText(getApplicationContext(), R.string.one_action, Toast.LENGTH_SHORT).show();
             }else {
                 task = task + "-";
             }
@@ -168,7 +164,7 @@ public class MainActivity extends AppCompatActivity {
                     (task.charAt(task.length() - 1) == '.')) {
                 task = task.substring(0, task.length() - 1) + "x";
             } else if (hasActionSign(task)){
-                Toast.makeText(getApplicationContext(), "Only one action a time, sorry", Toast.LENGTH_SHORT).show();
+                Toast.makeText(getApplicationContext(), R.string.one_action, Toast.LENGTH_SHORT).show();
             }else {
                 task = task + "x";
             }
@@ -184,7 +180,7 @@ public class MainActivity extends AppCompatActivity {
                     (task.charAt(task.length() - 1) == '.')) {
                 task = task.substring(0, task.length() - 1) + "÷";
             } else if (hasActionSign(task)){
-                Toast.makeText(getApplicationContext(), "Only one action a time, sorry", Toast.LENGTH_SHORT).show();
+                Toast.makeText(getApplicationContext(), R.string.one_action, Toast.LENGTH_SHORT).show();
             }else {
                 task = task + "÷";
             }
@@ -209,12 +205,11 @@ public class MainActivity extends AppCompatActivity {
 
         findViewById(R.id.b_equals).setOnClickListener(view -> {
             if (!hasActionSign(task)) {
-                Toast.makeText(getApplicationContext(), "There is no problem", Toast.LENGTH_SHORT).show();
+                Toast.makeText(getApplicationContext(), R.string.no_problem, Toast.LENGTH_SHORT).show();
             } else {
                 Calculus cl = new Calculus(task);
                 result = cl.Calculate(task);
                 displayResult.setText(result);
-                Toast.makeText(getApplicationContext(), "Calculating...", Toast.LENGTH_SHORT).show();
             }
             displayResult.setText(result);
         });
@@ -231,20 +226,8 @@ public class MainActivity extends AppCompatActivity {
         return (task.contains("+") || task.contains("-") || task.contains("x") || task.contains("÷"));
     }
 
-    private void saveNightModeState(boolean nightMode) {
-        SharedPreferences.Editor editor = sharedPreferences.edit();
-        editor.putBoolean(KEY_ISNIGHTMODE, nightMode);
-        editor.apply();
-    }
-    public void NightModeIsActivated(){
-        if(sharedPreferences.getBoolean(KEY_ISNIGHTMODE, false)){
-            nightSwitch.setChecked(true);
-            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
-        } else {
-            nightSwitch.setChecked(false);
-            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
-        }
-    }
+
+
 
     @Override
     protected void onSaveInstanceState(@NonNull Bundle instanceState) {
