+++ app/src/main/res/layout-night/activity_theme_chooser.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@drawable/background_night"
+    tools:context=".ThemeChooser">
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/b_return"
+        style="@style/PurpleStyleButtonsFunctionsPortrait"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:textSize="16sp"
+        android:text="Return" />
+
+    <androidx.appcompat.widget.SwitchCompat
+        android:id="@+id/switchNightMode"
+        style="@style/Displays"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingStart="24dp"
+        android:paddingEnd="24dp"
+        android:layout_marginTop="104dp"
+        android:background="@color/darkSlateBlue"
+        android:text="@string/night_mode"
+        android:textSize="@dimen/text_size_display_result"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:thumbTint="@color/lavenderBlush" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
