+++ app/src/main/res/values-ru/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+        <string name="night_mode">Ночной режим</string>
+        <string name="no_problem">Нет задачи</string>
+        <string name="one_action">Извините, только примеры в одно действие</string>
+        <string name="one_dot">У вашего числа уже есть точка</string>
+</resources>
